* java/lang/natClassLoader.cc (_Jv_RegisterClassHookDefault): 	Handle 'next' as version number.
2005-04-12  Aaron Luchko  <aluchko@redhat.com>

	* gnu/gcj/tools/gcj_dbtool/Main.java (addJar): Make 'length' an int.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gcjx-branch@98865 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libjava/gnu/gcj/tools/gcj_dbtool/Main.java
@@ -245,7 +245,7 @@ public class Main
 	if (classfile.getName().endsWith(".class"))
 	  {
 	    InputStream str = jar.getInputStream(classfile);
-	    long length = classfile.getSize();
+	    int length = (int) classfile.getSize();
 	    if (length == -1)
 	      throw new EOFException();
